created a separate module for firebase to use it in individual modules easier
